\name{QDG-package}

\alias{QDG-package}

\alias{QDG}

\docType{package}

\title{
Implements the QDG algorithm
}


\description{
Implements the QDG algorithm.
}


\details{
\tabular{ll}{
Package: \tab QDG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-06-17\cr
License: \tab GNU General Public License\cr
}



}
\author{
Elias Chaibub Neto and Brian S. Yandell

Maintainer: chaibub@stat.wisc.edu

}

\references{Chaibub Neto et al. (2008) Inferring causal phenotype networks from 
            segregating populations. Genetics 179: 1089-1100.}



\keyword{ package }


\seealso{

}

\examples{

}
